--- app/src/main/java/com/example/playlistmaker/data/api/ITunesApi.kt
@@ -1,11 +1,0 @@
-package com.example.playlistmaker.data.api
-
-import com.example.playlistmaker.data.dto.TracksResponse
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ITunesApi {
-    @GET("/search?entity=song")
-    fun search(@Query("term") text: String) : Call<TracksResponse>
-}
