--- app/src/main/java/com/example/playlistmaker/data/api/NetworkClient.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.data.api
-
-import com.example.playlistmaker.data.dto.Response
-
-interface NetworkClient {
-    fun doRequestApi(dto: Any): Response
-}
