--- app/src/main/java/com/example/playlistmaker/data/dto/TrackListAndResponse.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-import com.example.playlistmaker.domain.models.Track
-
-data class TrackListAndResponse(val trackList: List<Track>, val responseCode: Int)
