--- app/src/main/java/com/example/playlistmaker/data/dto/TracksResponse.kt
@@ -1,6 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-class TracksResponse(
-    val results: List<TracksDTO>,
-    resultCode: Int
-) : Response(resultCode)
