--- app/src/main/java/com/example/playlistmaker/data/dto/TracksSearchRequest.kt
@@ -1,3 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-data class TracksSearchRequest(val expression: String)
