--- app/src/main/java/com/example/playlistmaker/presentation/search/TrackViewHolder.kt
@@ -1,32 +1,0 @@
-package com.example.playlistmaker.presentation.search
-
-import android.view.View
-import android.widget.ImageView
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.example.playlistmaker.R
-import com.example.playlistmaker.domain.models.Track
-import java.text.SimpleDateFormat
-import java.util.Locale
-
-class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-    private val trackName: TextView = itemView.findViewById(R.id.track_name)
-    private val trackArtist: TextView = itemView.findViewById(R.id.track_artist)
-    private val trackTime: TextView = itemView.findViewById(R.id.track_time)
-    private val trackImage: ImageView = itemView.findViewById(R.id.track_image)
-
-    fun bind(model: Track) {
-        trackName.text = model.trackName
-        trackArtist.text = model.artistName
-        trackTime.text = SimpleDateFormat("mm:ss", Locale.getDefault()).format(model.trackTime)
-        Glide.with(itemView)
-            .load(model.artworkUrl100)
-            .centerCrop()
-            .transform(RoundedCorners(2))
-            .placeholder(R.drawable.img_placeholder)
-            .into(trackImage)
-    }
-
-}
