+++ app/src/main/java/com/example/playlistmaker/search/data/TrackDiffUtilCallback.kt
@@ -1,0 +1,15 @@
+package com.example.playlistmaker.search.data
+
+import androidx.recyclerview.widget.DiffUtil
+import com.example.playlistmaker.search.domain.models.Track
+
+class TrackDiffUtilCallback : DiffUtil.ItemCallback<Track>() {
+
+    override fun areItemsTheSame(oldItem: Track, newItem: Track): Boolean {
+        return oldItem.trackId == newItem.trackId
+    }
+
+    override fun areContentsTheSame(oldItem: Track, newItem: Track): Boolean {
+        return oldItem == newItem
+    }
+}
