+++ app/src/main/java/com/example/playlistmaker/search/data/api/NetworkClient.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.search.data.api
+
+import com.example.playlistmaker.search.data.dto.Response
+
+interface NetworkClient {
+    fun doRequestApi(dto: Any): Response
+}
