+++ app/src/main/java/com/example/playlistmaker/search/data/dto/TrackListAndResponse.kt
@@ -1,0 +1,5 @@
+package com.example.playlistmaker.search.data.dto
+
+import com.example.playlistmaker.search.domain.models.Track
+
+data class TrackListAndResponse(val trackList: List<Track>, val responseCode: Int)
