+++ app/src/main/java/com/example/playlistmaker/search/data/dto/TracksResponse.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.search.data.dto
+
+class TracksResponse(
+    val results: List<TracksDTO>,
+    resultCode: Int
+) : Response(resultCode)
