+++ app/src/main/java/com/example/playlistmaker/search/domain/TracksInteractor.kt
@@ -1,0 +1,11 @@
+package com.example.playlistmaker.search.domain
+
+import com.example.playlistmaker.search.domain.models.Track
+
+interface TracksInteractor {
+    fun doRequest(expression: String, consumer: TracksConsumer)
+
+    interface TracksConsumer {
+        fun consume(findTracks: List<Track>, response: Int)
+    }
+}
