+++ app/src/main/java/com/example/playlistmaker/search/domain/TracksRepository.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.search.domain
+
+import com.example.playlistmaker.search.data.dto.TrackListAndResponse
+
+interface TracksRepository {
+//    fun refillTrackList(expression: String): List<Track>
+    fun refillTrackList(expression: String): TrackListAndResponse
+}
