+++ app/src/main/res/layout/activity_media.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".presentation.media.MediaActivity"
+    tools:context=".media.ui.MediaActivity"
     tools:ignore="UseCompoundDrawables">
 
     <TextView
